[NAYB-46] refactor : 엔티티 경로 수정에 의한 import 경로 수정
