+++ src/test/java/com/prgrms/nabmart/global/fixture/ItemFixture.java
@@ -1,7 +1,7 @@
 package com.prgrms.nabmart.global.fixture;
 
-import com.prgrms.nabmart.domain.category.domain.MainCategory;
-import com.prgrms.nabmart.domain.category.domain.SubCategory;
+import com.prgrms.nabmart.domain.category.MainCategory;
+import com.prgrms.nabmart.domain.category.SubCategory;
 import com.prgrms.nabmart.domain.item.domain.Item;
 
 public final class ItemFixture {
